requirements.txt
asgiref==3.11.0
click==8.3.1
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-rest-framework==0.1.0
django-types==0.22.0
djangorestframework==3.16.1
djangorestframework-types==0.9.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.11
psycopgbinary==0.0.1
sqlparse==0.5.4
types-psycopg2==2.9.21.20251012
uvicorn==0.38.0
